+++ src/features/manage/components/manage-messages/manage-messages.module.scss
@@ -1,6 +1,6 @@
 @import 'styles/convert.scss';
 
-.messages {
+.manageMessages {
   display: flex;
   flex-direction: column;
   width: rem(750px);
@@ -8,22 +8,22 @@
   margin: rem(32px) 0 rem(68px) 0;
 }
 
-.messagesTextarea {
+.manageMessagesTextarea {
   width: 100%;
   height: rem(150px);
 }
 
-.messagesButton {
+.manageMessagesButton {
   width: 100%;
   display: flex;
   justify-content: flex-end;
 }
 
-.messagesButtonStyle {
+.manageMessagesButtonStyle {
   width: rem(165px);
 }
 
-.messagesButtonDisable {
+.manageMessagesButtonDisable {
   width: rem(165px);
   pointer-events: none;
 }
