--- src/features/manage/components/manage-photos/manage-photos.tsx
@@ -1,3 +1,0 @@
-export const ManagePhotos = () => {
-  return <p>manage-photos</p>
-}
