+++ src/features/manage/constants/manage-constants.ts
@@ -1,11 +1,23 @@
-export const MANAGE_LABELS = {
-  ADD_NEW: 'Add new',
-  DETAILS: 'Details',
-  PHOTO: 'Photos & videos',
-  DESCRIPTION: 'Description',
-  CONTACT: 'Contact info',
+export const PROPRETY_ALT_IMG = {
+  IMAGE_HOME: 'Home Photo',
+  SHOW: 'Show all photo',
+  AVATAR: 'User Photo',
+  SHARE: 'Share Icon',
+  FAVORITE: 'Favorite Icon',
 }
 
-export const MANAGE_ALT_IMG = {
-  ADD: 'Add new Icon',
+export const PROPRETY_BUTTON = {
+  ALL_IMG: 'Show all photos',
+  PURCHASE: 'Purchase',
+  DESCRIPTION: 'Description',
+  SHARE: 'Share',
+  SEND: 'Send',
+}
+
+export const PROPRETY_LABELS = {
+  MESSAGES_LABELS: 'Message the seller',
+}
+
+export const PROPRETY_PLACEHOLDER = {
+  TEXTAREA: 'Messages..',
 }
