+++ src/pages/page-all-components.tsx
@@ -5,7 +5,7 @@ import { ShowroomPropertyList } from 'features/showroom/components/showroom-prop
 import { ShowroomViewSwitch } from 'features/showroom/components/showroom-view-switch/showroom-view-switch'
 import { SHOWROOM_ACTIVE_VIEW } from 'features/showroom/constants/showroom-constants'
 import { viewType } from 'features/showroom/models/showroom-models'
-import { ResultsMessages } from 'features/results/components/proprety-messages/results-messages'
+import { ManageMessages } from 'features/manage/components/manage-messages/manage-messages'
 
 export const PageAllComponents = () => {
   const [activeView, setActiveView] = useState<viewType>(SHOWROOM_ACTIVE_VIEW.LIST as viewType)
@@ -15,7 +15,7 @@ export const PageAllComponents = () => {
       <ShowroomAllComponents />
       <ShowroomViewSwitch activeView={activeView} switchView={(view) => setActiveView(view)} />
       <ShowroomPropertyList title={'Small'} />
-      <ResultsMessages />
+      <ManageMessages />
     </div>
   )
 }
